requirements.txt
streamlit>=1.32.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-chroma>=0.1.0
langchain-text-splitters>=0.2.0
chromadb>=0.5.0
openai>=1.25.0
pypdf>=4.2.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
pytest>=8.0.0
